Added evrDiag utility to read registers.
Rewrote driver to *not* map registers in shared mode.  Most registers can be read with an ioctl, except the
event queue is explicitly disabled.  Added another ioctl to program pulses, and added locking so that only one
shared connection can program a particular pulse.  Fixed a lot of unsigned long vs. EpicsUInt32 nonsense, and
wrote a little warning about it.



git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/event2/current@13299 ececa60f-9f63-0410-97c7-ea05888eb260